feat!: remove supportsInterface fn
Using vyper we get a fn sig that doesn't match with the ERC165 standard,
meaning if any contract were to call the function on chain it would
fail. Therefore it's pointless to include it in the source code,
especially since users who can read the source, will be able to identify
the standards which this contract abides by.